Update patch set 1
Patch Set 1:

When I ran release.sh it updated the pom.xml while building the release war so the embedded version number was correct, but it never made a commit in that state.

I was under the impression that if you did "git checkout v2.7" then all the pom.xmls should have the actual non-snapshot version number; is that not the case?

Patch-set: 1