Create review note also when newObjectId already present in another branch
Using the %base option of the refs/for/master ref one can push a commit
for review even when it is already known to Gerrit (already present in
another branch). Since we used to mark as uninteresting the tips of all
branches, the tree walk used to identify commits for which to create review
notes wouldn't find such a commit.

This change fixes only the easy case when the newObjectId repesents a normal
(non-merge) commit and when the oldObjectId is the parent of the newObjectId.

Change-Id: Ide86cde5a8359b8764a9e1984d29865b604059ec