ExportReviewNotes: Load changes from notedb if enabled
We keep the code for loading the caches from the database since it
allows us to load all changes as a batch, which is not possible via
ChangeNotes.Factory.

Also use Multimap<Project.NameKey, Change> instead of
Map<Project.NameKey, List<Change>>. For this some parameter types had
to be changed from List to Collection.

Change-Id: Ie9b936cbedc333c0d7697af706900187a34e7060
Signed-off-by: Edwin Kempin <ekempin@google.com>